Move downloadCSV function to its own file
